Move to virtual cordapp
